requirements.txt
PyGithub==2.3.0
python-dotenv==1.0.1
nicegui==1.4.22
discord.py==2.3.2
requests==2.32.3
